Started creating real game logic.

+++ src/messages.rs
@@ -5,21 +5,25 @@ pub enum Messages {
     CompleteDirective,
 }
 
+#[derive(Copy, Clone)]
 pub enum Interface {
     Eigenthrottle,
 }
 
+#[derive(Copy, Clone)]
 pub enum Value {
     Enable,
     Disable,
 }
 
+#[derive(Copy, Clone)]
 pub struct Action {
     pub interface: Interface,
     pub value: Value,
 }
 
+#[derive(Copy, Clone)]
 pub struct Directive {
     pub action: Action,
-    pub time_ms: u16,
+    pub time_ms: u32,
 }
